requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
Werkzeug==3.0.1
certifi==2023.11.17